End Test for commit e4dad005190648f8ce2a9109446ed623e4b500f9 in main branch. - Auth View-Model login Test cases     * Accepted Login Test.     * Invalid Email exception.     * Invalid Password exception {Less than 8 characters}.     * Invalid Password exception {Regular exception error}.     * Network exception error {Error in email or password}.
